Implementa a classe PdoDatabase e interfaces de driver, adicionando suporte para MySQL, PostgreSQL e SQLite
